refactor(libsy): address review comments
- Remove switchyard-translation dev dep; rewrite tool_signals tests to
  construct LlmRequest directly
- Move stage_router integration tests into stage.rs #[cfg(test)]
- Trim paragraph comments in fall_through.rs
