+++ crates/libsy/src/algorithms/util/tool_signals.rs
@@ -597,33 +597,50 @@ fn has_nonzero_failure_count(lower: &str) -> bool {
 mod tests {
     use super::*;
     use serde_json::json;
-    use switchyard_protocol::{Metadata, WireFormat};
+    use switchyard_protocol::{ContentBlock, LlmRequest, Message, Role, ToolCall, ToolResult};
 
-    /// Decodes `body` the way an endpoint would, so these tests assert on the
-    /// bodies a real client sends rather than on hand-built internal shapes.
-    fn request_with(wire_format: WireFormat, body: Value) -> Request {
-        let llm_request = switchyard_translation::decode_request(wire_format, &body)
-            .unwrap_or_else(|error| panic!("fixture is not valid {wire_format}: {error}"));
+    fn with_messages(messages: Vec<Message>) -> Request {
         Request {
-            llm_request,
-            raw_request: Some(body),
-            metadata: Some(Metadata {
-                wire_format: Some(wire_format),
-                ..Default::default()
-            }),
+            llm_request: LlmRequest { messages, ..LlmRequest::default() },
+            raw_request: None,
+            metadata: None,
         }
     }
 
-    fn openai_chat_request(body: Value) -> Request {
-        request_with(WireFormat::OpenAiChat, body)
+    // assistant message with a single named tool call
+    fn tc(name: &str) -> Message {
+        Message {
+            role: Role::Assistant,
+            content: vec![ContentBlock::ToolCall(ToolCall {
+                id: String::new(),
+                name: name.to_string(),
+                arguments: json!({}),
+            })],
+        }
     }
 
-    fn anthropic_request(body: Value) -> Request {
-        request_with(WireFormat::AnthropicMessages, body)
+    // assistant Bash message carrying `command`
+    fn bash(command: &str) -> Message {
+        Message {
+            role: Role::Assistant,
+            content: vec![ContentBlock::ToolCall(ToolCall {
+                id: String::new(),
+                name: "Bash".to_string(),
+                arguments: json!({"command": command}),
+            })],
+        }
     }
 
-    fn openai_responses_request(body: Value) -> Request {
-        request_with(WireFormat::OpenAiResponses, body)
+    // a tool result message (goes in a user-role message, as in Anthropic's normalised form)
+    fn tr(text: &str) -> Message {
+        Message {
+            role: Role::User,
+            content: vec![ContentBlock::ToolResult(ToolResult {
+                tool_call_id: String::new(),
+                content: vec![ContentBlock::Text { text: text.to_string() }],
+                is_error: None,
+            })],
+        }
     }
 
     #[test]
@@ -705,14 +722,11 @@ mod tests {
     #[test]
     fn severity_is_windowed_over_recent_results() {
         // An error two results back, then two clean results.
-        let request = openai_chat_request(json!({
-            "messages": [
-                {"role": "tool", "tool_call_id": "1",
-                 "content": "Traceback (most recent call last):\n  ValueError"},
-                {"role": "tool", "tool_call_id": "2", "content": "ok"},
-                {"role": "tool", "tool_call_id": "3", "content": "ok"},
-            ]
-        }));
+        let request = with_messages(vec![
+            tr("Traceback (most recent call last):\n  ValueError"),
+            tr("ok"),
+            tr("ok"),
+        ]);
         // window covers the error → severity persists (max over the window)
         assert_eq!(extract_tool_signals_with_window(&request, 3).severity, HARD);
         // window of 1 sees only the last (clean) result → severity has decayed out
@@ -721,13 +735,11 @@ mod tests {
 
     #[test]
     fn extract_openai_chat_tool_results() {
-        let request = openai_chat_request(json!({
-            "messages": [
-                {"role": "user", "content": "do something"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Edit"}}]},
-                {"role": "tool", "tool_call_id": "1", "content": "Traceback (most recent call last):\n  ValueError"},
-            ]
-        }));
+        let request = with_messages(vec![
+            Message::text(Role::User, "do something"),
+            tc("Edit"),
+            tr("Traceback (most recent call last):\n  ValueError"),
+        ]);
         let sig = ToolSignals::from_request(&request, None);
         assert_eq!(sig.severity, HARD);
         assert_eq!(sig.edit_count, 1);
@@ -736,27 +748,19 @@ mod tests {
 
     #[test]
     fn extract_anthropic_tool_results() {
-        let request = anthropic_request(json!({
-            "messages": [
-                {"role": "user", "content": [
-                    {"type": "tool_result", "tool_use_id": "1",
-                     "content": "Traceback (most recent call last):\n  ValueError"}
-                ]},
-            ]
-        }));
+        let request = with_messages(vec![
+            tr("Traceback (most recent call last):\n  ValueError"),
+        ]);
         let sig = ToolSignals::from_request(&request, None);
         assert_eq!(sig.severity, HARD);
     }
 
     #[test]
     fn extract_responses_api_tool_results() {
-        let request = openai_responses_request(json!({
-            "input": [
-                {"type": "function_call", "name": "Write"},
-                {"type": "function_call_output", "call_id": "1",
-                 "output": "file written successfully"},
-            ]
-        }));
+        let request = with_messages(vec![
+            tc("Write"),
+            tr("file written successfully"),
+        ]);
         let sig = ToolSignals::from_request(&request, None);
         assert_eq!(sig.severity, 0.0);
         assert_eq!(sig.write_count, 1);
@@ -766,22 +770,14 @@ mod tests {
     fn recent_window_counts_only_last_default_window_tool_calls() {
         // 5 writes + 1 edit at the end → the default window (3) should see
         // the last 3 calls: 1 edit + 2 writes (not all 6 calls).
-        let request = openai_chat_request(json!({
-            "messages": [
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Write"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Write"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Write"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Write"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Write"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Edit"}}]},
-                {"role": "tool", "content": "ok"},
-            ]
-        }));
+        let request = with_messages(vec![
+            tc("Write"), tr("ok"),
+            tc("Write"), tr("ok"),
+            tc("Write"), tr("ok"),
+            tc("Write"), tr("ok"),
+            tc("Write"), tr("ok"),
+            tc("Edit"),  tr("ok"),
+        ]);
         let sig = ToolSignals::from_request(&request, None);
         assert_eq!(sig.write_count, 5);
         assert_eq!(sig.edit_count, 1);
@@ -794,22 +790,14 @@ mod tests {
         // Same six tool calls (1 edit at the end, 5 writes before).
         // With recent_window=3 → recent_writes=2, recent_edits=1.
         // With recent_window=6 → recent_writes=5, recent_edits=1 (all calls).
-        let request = openai_chat_request(json!({
-            "messages": [
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Write"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Write"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Write"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Write"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Write"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Edit"}}]},
-                {"role": "tool", "content": "ok"},
-            ]
-        }));
+        let request = with_messages(vec![
+            tc("Write"), tr("ok"),
+            tc("Write"), tr("ok"),
+            tc("Write"), tr("ok"),
+            tc("Write"), tr("ok"),
+            tc("Write"), tr("ok"),
+            tc("Edit"),  tr("ok"),
+        ]);
         let narrow = extract_tool_signals_with_window(&request, 3);
         assert_eq!(narrow.recent_write_count, 2);
         assert_eq!(narrow.recent_edit_count, 1);
@@ -822,39 +810,28 @@ mod tests {
     #[test]
     fn compaction_marker_sets_compacted() {
         // The compaction summary is a user message carrying Claude Code's preamble.
-        let request = openai_chat_request(json!({
-            "messages": [
-                {"role": "user", "content": "This session is being continued from a previous conversation that ran out of context."},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Bash", "arguments": "{\"command\": \"ls\"}"}}]},
-            ]
-        }));
+        let request = with_messages(vec![
+            Message::text(Role::User, "This session is being continued from a previous conversation that ran out of context."),
+            bash("ls"),
+        ]);
         assert!(ToolSignals::from_request(&request, None).compacted);
     }
 
     #[test]
     fn no_compaction_marker_stays_uncompacted() {
-        let request = openai_chat_request(json!({
-            "messages": [
-                {"role": "user", "content": "Write a script that parses the log file."},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Bash", "arguments": "{\"command\": \"ls\"}"}}]},
-            ]
-        }));
+        let request = with_messages(vec![
+            Message::text(Role::User, "Write a script that parses the log file."),
+            bash("ls"),
+        ]);
         assert!(!ToolSignals::from_request(&request, None).compacted);
     }
 
     #[test]
     fn bash_heredoc_counts_as_write() {
         // Claude Code's pattern on TB 2.0 — write a scratch file via heredoc.
-        let request = openai_chat_request(json!({
-            "messages": [
-                {"role": "assistant", "tool_calls": [{
-                    "function": {
-                        "name": "Bash",
-                        "arguments": "{\"command\": \"cat > /tmp/test.py <<'EOF'\\nprint(1)\\nEOF\"}"
-                    }
-                }]},
-            ]
-        }));
+        let request = with_messages(vec![
+            bash("cat > /tmp/test.py <<'EOF'\nprint(1)\nEOF"),
+        ]);
         let sig = ToolSignals::from_request(&request, None);
         assert_eq!(
             sig.write_count, 1,
@@ -865,16 +842,9 @@ mod tests {
 
     #[test]
     fn bash_sed_inplace_counts_as_edit() {
-        let request = openai_chat_request(json!({
-            "messages": [
-                {"role": "assistant", "tool_calls": [{
-                    "function": {
-                        "name": "Bash",
-                        "arguments": "{\"command\": \"sed -i 's/foo/bar/g' /app/file.py\"}"
-                    }
-                }]},
-            ]
-        }));
+        let request = with_messages(vec![
+            bash("sed -i 's/foo/bar/g' /app/file.py"),
+        ]);
         let sig = ToolSignals::from_request(&request, None);
         assert_eq!(
             sig.edit_count, 1,
@@ -886,16 +856,10 @@ mod tests {
     #[test]
     fn bash_non_mutating_does_not_count() {
         // ls, cat, grep — should not increment either counter.
-        let request = openai_chat_request(json!({
-            "messages": [
-                {"role": "assistant", "tool_calls": [{
-                    "function": {"name": "Bash", "arguments": "{\"command\": \"ls -la /app\"}"}
-                }]},
-                {"role": "assistant", "tool_calls": [{
-                    "function": {"name": "Bash", "arguments": "{\"command\": \"cat /app/main.py\"}"}
-                }]},
-            ]
-        }));
+        let request = with_messages(vec![
+            bash("ls -la /app"),
+            bash("cat /app/main.py"),
+        ]);
         let sig = ToolSignals::from_request(&request, None);
         assert_eq!(sig.write_count, 0);
         assert_eq!(sig.edit_count, 0);
@@ -958,14 +922,9 @@ mod tests {
     #[test]
     fn anthropic_bash_heredoc_extracts_command() {
         // Anthropic format: tool_use.input is an object, not a JSON string.
-        let request = anthropic_request(json!({
-            "messages": [
-                {"role": "assistant", "content": [
-                    {"type": "tool_use", "name": "Bash",
-                     "input": {"command": "cat > /tmp/foo.txt << 'EOF'\nhi\nEOF"}}
-                ]},
-            ]
-        }));
+        let request = with_messages(vec![
+            bash("cat > /tmp/foo.txt << 'EOF'\nhi\nEOF"),
+        ]);
         let sig = ToolSignals::from_request(&request, None);
         assert_eq!(
             sig.write_count, 1,
@@ -975,11 +934,9 @@ mod tests {
 
     #[test]
     fn recent_window_falls_back_to_full_history_when_short() {
-        let request = openai_chat_request(json!({
-            "messages": [
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Write"}}]},
-            ]
-        }));
+        let request = with_messages(vec![
+            tc("Write"),
+        ]);
         let sig = ToolSignals::from_request(&request, None);
         assert_eq!(sig.recent_write_count, 1);
         assert_eq!(sig.recent_edit_count, 0);
@@ -987,12 +944,10 @@ mod tests {
 
     #[test]
     fn clean_tool_result_has_zero_severity_and_non_empty_streak() {
-        let request = openai_chat_request(json!({
-            "messages": [
-                {"role": "tool", "tool_call_id": "1", "content": "output ok"},
-                {"role": "tool", "tool_call_id": "2", "content": "another ok"},
-            ]
-        }));
+        let request = with_messages(vec![
+            tr("output ok"),
+            tr("another ok"),
+        ]);
         let sig = ToolSignals::from_request(&request, None);
         assert_eq!(sig.severity, 0.0);
         assert_eq!(sig.no_error_streak, 2);
@@ -1066,25 +1021,13 @@ mod tests {
     #[test]
     fn pure_bash_streak_counts_trailing_other() {
         // 5 trailing non-classified Bash calls → streak == 5.
-        let request = openai_chat_request(json!({
-            "messages": [
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Bash",
-                    "arguments": "{\"command\": \"make\"}"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Bash",
-                    "arguments": "{\"command\": \"./configure\"}"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Bash",
-                    "arguments": "{\"command\": \"make install\"}"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Bash",
-                    "arguments": "{\"command\": \"./run.sh\"}"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Bash",
-                    "arguments": "{\"command\": \"./test\"}"}}]},
-                {"role": "tool", "content": "ok"},
-            ]
-        }));
+        let request = with_messages(vec![
+            bash("make"),         tr("ok"),
+            bash("./configure"),  tr("ok"),
+            bash("make install"), tr("ok"),
+            bash("./run.sh"),     tr("ok"),
+            bash("./test"),       tr("ok"),
+        ]);
         let sig = ToolSignals::from_request(&request, None);
         assert_eq!(sig.pure_bash_streak, 5);
         assert_eq!(sig.write_count, 0);
@@ -1093,15 +1036,10 @@ mod tests {
 
     #[test]
     fn pure_bash_streak_resets_on_write() {
-        let request = openai_chat_request(json!({
-            "messages": [
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Bash",
-                    "arguments": "{\"command\": \"make\"}"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Write"}}]},
-                {"role": "tool", "content": "ok"},
-            ]
-        }));
+        let request = with_messages(vec![
+            bash("make"), tr("ok"),
+            tc("Write"),  tr("ok"),
+        ]);
         let sig = ToolSignals::from_request(&request, None);
         assert_eq!(sig.pure_bash_streak, 0);
         assert_eq!(sig.write_count, 1);
@@ -1110,19 +1048,12 @@ mod tests {
     #[test]
     fn recent_window_tracks_todowrite_and_read() {
         // Final 3 tool calls: TodoWrite, Read, TodoWrite.
-        let request = openai_chat_request(json!({
-            "messages": [
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Bash",
-                    "arguments": "{\"command\": \"make\"}"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "TodoWrite"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "Read"}}]},
-                {"role": "tool", "content": "ok"},
-                {"role": "assistant", "tool_calls": [{"function": {"name": "TodoWrite"}}]},
-                {"role": "tool", "content": "ok"},
-            ]
-        }));
+        let request = with_messages(vec![
+            bash("make"),      tr("ok"),
+            tc("TodoWrite"),   tr("ok"),
+            tc("Read"),        tr("ok"),
+            tc("TodoWrite"),   tr("ok"),
+        ]);
         let sig = ToolSignals::from_request(&request, None);
         assert_eq!(sig.todowrite_count, 2);
         assert_eq!(sig.recent_todowrite_count, 2);
